// src/types/value-item.ts
import { Value } from '@/fixtures/valuesItems/valuesItems';

export interface ValueItemProps {
  item: Value;
}

export interface ClassNames {
  classNames: {
    item: string;
    itemContainer: string;
    itemIcon: string;
    itemTitles: string;
    itemTitle: string;
    itemSubTitle: string;
  };
}